+++ src/helpers/define-api-host.ts
@@ -14,7 +14,7 @@ const defineApiHost: TDefineApiHost = (
   switch (Number(chainId)) {
     case EChains.polygon: return `${apiHostActual}/polygon`
     case EChains.base: return `${apiHostActual}/base`
-    case EChains.mumbai: return `${apiHostActual}/mumbai`
+    case EChains.sepolia: return `${apiHostActual}/sepolia`
     case EChains.baseGoerli: return `${apiHostActual}/base-goerli`
     case EChains.arbitrum: return `${apiHostActual}/arbitrum`
     case EChains.optimism: return `${apiHostActual}/optimism`