+++ src/types/token-address/index.ts
@@ -3,9 +3,7 @@ export enum ETokenAddress {
   usdcPolygon = '0x3c499c542cef5e3811e1192ce70d8cc03d5c3359',
   usdcBase = '0x833589fcd6edb6e08f4c7c32d4f71b54bda02913',
   usdcBaseGoerli = '0xf175520c52418dfe19c8098071a252da48cd1c19',
-  usdcMumbai = '0x0fa8781a83e46826621b3bc094ea2a0212e71b23',
-
-
+  usdcSepolia = '0x1c7d4b196cb0c7b01d743fbc6116a902379c7238',
   usdcArbitrum = '0xaf88d065e77c8cc2239327c5edb3a432268e5831',
   usdcOptimism = '0x0b2c639c533813f4aa9d7837caf62653d097ff85',
   usdcAvalanche = '0xb97ef9ef8734c71904d8002f8b6bc66dd9c48a6e'